+++ lib/saf_util_method_channel.dart
@@ -136,8 +136,8 @@ class MethodChannelSafUtil extends SafUtilPlatform {
   @override
   Future<bool> saveThumbnailToFile({
     required String uri,
-    required double width,
-    required double height,
+    required int width,
+    required int height,
     required String destPath,
     String? format,
     int? quality,
